Refactor Repository
